+++ src/App.js
@@ -2,7 +2,8 @@ import "./App.css";
 
 //컴포넌트 요소
 import Header from "./components/Header/Header";
-import MemberLogin from "./pages/MemberLogin/MemberLogin";
+// import MemberLogin from "./pages/MemberLogin/MemberLogin";
+import Member from "./pages/Member/Member";
 import Footer from "./components/Footer/Footer";
 import Main from "./pages/Main/Main";
 import Denied from "./components/Denied/Denied";
@@ -39,7 +40,7 @@ function App() {
             <Header></Header>
             <Routes>
               <Route path="/" element={<Main />} />
-              <Route path="/login/*" element={<MemberLogin />}></Route>
+              <Route path="/member/*" element={<Member />}></Route>
               <Route path="/party/*" element={<PartyCreate />}></Route>
               <Route path="/Mypage/*" element={<Mypage />}></Route>
               <Route path="/denied" element={<Denied />}></Route>
