+++ src/pages/Member/AccountRegistration/AccountRegistration.js
@@ -1,0 +1,5 @@
+import style from "./AccountRegistration.module.css";
+
+const AccountRegistration = () => {};
+
+export default AccountRegistration;
