+++ src/pages/Member/SignUp/SignUp.module.css
@@ -2,7 +2,7 @@
   margin: 100px auto 150px auto;
   max-width: 1300px;
 }
-.title {
+/* .title {
   font-size: 40px;
   font-weight: 700;
   text-align: center;
@@ -53,7 +53,7 @@
 
 .stepBox__arrow > i {
   color: #7d7d7d;
-}
+} */
 
 .agrreBox,
 .signupBox,
@@ -62,7 +62,7 @@
   margin-top: 30px;
 }
 
-.useAgree,
+/* .useAgree,
 .privacyAgree {
   margin-top: 30px;
 }
@@ -89,10 +89,10 @@
 .useDetail::-webkit-scrollbar-thumb {
   background-color: #d1d2d8;
   border-radius: 10px;
-  border: 7px solid transparent; /* 스크롤을 적용할 영역 색깔과 border 색상을 똑같이 맞춘다 */
+  border: 7px solid transparent; 
 }
 .useDetail::-webkit-scrollbar-track {
-  background-color: #d1d2d850; /* 스크롤바 뒷 배경을 투명 처리한다 */
+  background-color: #d1d2d850; 
 }
 .privacyAgree::-webkit-scrollbar {
   width: 10px;
@@ -100,11 +100,10 @@
 .privacyAgree::-webkit-scrollbar-thumb {
   background-color: #d1d2d8;
   border-radius: 10px;
-  border: 7px solid transparent; /* 스크롤을 적용할 영역 색깔과 border 색상을 똑같이 맞춘다 */
+  border: 7px solid transparent; 
 }
 .privacyAgree::-webkit-scrollbar-track {
-  background-color: #d1d2d850; /* 스크롤바 뒷 배경을 투명 처리한다 */
-}
+  background-color: #d1d2d850; 
 
 .article:not(:first-child) {
   margin-top: 50px;
@@ -120,7 +119,7 @@
 
 .article__text:not(:first-child) {
   margin-top: 20px;
-}
+} */
 .stepBtns {
   text-align: center;
   margin-top: 50px;
@@ -128,22 +127,22 @@
 .stepBtns button:last-child {
   margin-left: 20px;
 }
-.signupTitle {
+/* .signupTitle {
   padding: 0 0 7px 24px;
   font-size: 20px;
   font-weight: 700;
 }
 .signupTitle:not(:first-child) {
   margin-top: 100px;
-}
-.basicInfo,
+} */
+/* .basicInfo,
 .myInfo,
 .etcInfo {
   border: 1px solid #676d7c50;
   border-radius: 20px;
   padding: 16px 24px;
-}
-.basicInfo__line,
+} */
+/* .basicInfo__line,
 .myInfo__line,
 .etcInfo__line {
   display: flex;
@@ -160,6 +159,7 @@
 }
 .signup__inputBox {
   width: calc(100% - 120px);
+  overflow: hidden;
 }
 .signup__inputBox input {
   padding: 10px;
@@ -175,7 +175,7 @@
   padding-left: 10px;
   overflow: hidden;
   text-overflow: ellipsis;
-  /* white-space: nowrap; */
+  white-space: nowrap;
 }
 .myInfo__line:nth-child(4) .signup__inputBox,
 .myInfo__line:last-child .signup__inputBox {
@@ -197,7 +197,7 @@
   height: 40px;
   display: flex;
   align-items: center;
-}
+} */
 .currentBtn {
   display: flex;
 }
@@ -267,17 +267,17 @@
     margin: 100px auto 100px auto;
   }
 
-  .stepBox__step {
+  /* .stepBox__step {
     padding: 10px;
     width: calc((100% - 30px) / 3);
   }
   .stepBox__arrow {
     width: 15px;
-  }
+  } */
   .stepBtns button {
     width: 150px !important;
   }
-  .myInfo__line:nth-child(4) .signup__inputBox input,
+  /* .myInfo__line:nth-child(4) .signup__inputBox input,
   .myInfo__line:last-child .signup__inputBox input {
     width: 100%;
   }
@@ -296,5 +296,5 @@
   .myInfo__line:nth-child(4) .signup__chk button,
   .myInfo__line:last-child .signup__chk button {
     margin: 0 10px 0 0;
-  }
+  } */
 }
