+++ src/pages/Member/SignUp/Step1/Step1.module.css
@@ -1,0 +1,76 @@
+.signupBox {
+  margin: 100px auto 150px auto;
+  max-width: 1300px;
+}
+.agrreBox {
+  padding: 0 35px;
+  margin-top: 30px;
+}
+.useAgree,
+.privacyAgree {
+  margin-top: 30px;
+}
+.agrreBox input,
+.agrreBox label {
+  cursor: pointer;
+}
+.agree__span {
+  color: #ff4500;
+  margin-left: 10px;
+}
+.useDetail {
+  margin-top: 20px;
+  border: 1px solid #d1d2d850;
+  padding: 20px;
+  background-color: #d1d2d850;
+  height: 400px;
+  overflow-y: auto;
+}
+.useDetail::-webkit-scrollbar {
+  width: 10px;
+}
+.useDetail::-webkit-scrollbar-thumb {
+  background-color: #d1d2d8;
+  border-radius: 10px;
+  border: 7px solid transparent; /* 스크롤을 적용할 영역 색깔과 border 색상을 똑같이 맞춘다 */
+}
+.useDetail::-webkit-scrollbar-track {
+  background-color: #d1d2d850; /* 스크롤바 뒷 배경을 투명 처리한다 */
+}
+.privacyAgree::-webkit-scrollbar {
+  width: 10px;
+}
+.privacyAgree::-webkit-scrollbar-thumb {
+  background-color: #d1d2d8;
+  border-radius: 10px;
+  border: 7px solid transparent; /* 스크롤을 적용할 영역 색깔과 border 색상을 똑같이 맞춘다 */
+}
+.privacyAgree::-webkit-scrollbar-track {
+  background-color: #d1d2d850; /* 스크롤바 뒷 배경을 투명 처리한다 */
+}
+.article:not(:first-child) {
+  margin-top: 50px;
+}
+.article__title {
+  font-weight: 700;
+  margin-bottom: 20px;
+  font-size: 16px;
+}
+.article__text {
+  font-size: 14px;
+}
+.article__text:not(:first-child) {
+  margin-top: 20px;
+}
+.stepBtns {
+  text-align: center;
+  margin-top: 50px;
+}
+.stepBtns button:last-child {
+  margin-left: 20px;
+}
+@media (max-width: 786px) {
+  .stepBtns button {
+    width: 150px !important;
+  }
+}
